Rewrote the the way to style the custom dropdown component.
Since the whole structure of the dropdown component is separated, hence DCSS(Stands for `Dropdown Cascading Style Sheets`) is implemented to style the dropdown component. I have tried my best to make the styling style as close as the original CSS and the `DCSSProperties` interface is fully customizable due to personal preference. At the meanwhile, the users are allowed to use the original CSS to style the whole component(Make sure to use when the CSS property is "inheritable" by default to the child), eg. 'position', 'top', 'fontFamily', 'fontSize', 'textAlign' and more...